Update from  original

+++ src/Umbraco.Core/PluginManager.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Reflection;
 using System.Text;
 using System.Threading;
+using System.Web;
 using System.Web.Compilation;
 using Umbraco.Core.Cache;
 using Umbraco.Core.IO;
@@ -15,6 +16,7 @@ using Umbraco.Core.Persistence.SqlSyntax;
 using Umbraco.Core.Profiling;
 using Umbraco.Core.PropertyEditors;
 using umbraco.interfaces;
+using Umbraco.Core.Configuration;
 using File = System.IO.File;
 
 namespace Umbraco.Core
@@ -44,8 +46,8 @@ namespace Umbraco.Core
         private readonly object _typesLock = new object();
         private readonly Dictionary<TypeListKey, TypeList> _types = new Dictionary<TypeListKey, TypeList>();
 
-        private long _cachedAssembliesHash = -1;
-        private long _currentAssembliesHash = -1;
+        private string _cachedAssembliesHash = null;
+        private string _currentAssembliesHash = null;
         private IEnumerable<Assembly> _assemblies;
         private bool _reportedChange;
 
@@ -75,16 +77,17 @@ namespace Umbraco.Core
 
             if (detectChanges)
             {
-                //first check if the cached hash is 0, if it is then we ne
+                //first check if the cached hash is string.Empty, if it is then we need
                 //do the check if they've changed
-                RequiresRescanning = (CachedAssembliesHash != CurrentAssembliesHash) || CachedAssembliesHash == 0;
+                RequiresRescanning = (CachedAssembliesHash != CurrentAssembliesHash) || CachedAssembliesHash == string.Empty;
                 //if they have changed, we need to write the new file
                 if (RequiresRescanning)
                 {
                     // if the hash has changed, clear out the persisted list no matter what, this will force
                     // rescanning of all plugin types including lazy ones.
                     // http://issues.umbraco.org/issue/U4-4789
-                    File.Delete(pluginListFile);
+                    if(File.Exists(pluginListFile))
+                        File.Delete(pluginListFile);
 
                     WriteCachePluginsHash();
                 }
@@ -94,7 +97,8 @@ namespace Umbraco.Core
                 // if the hash has changed, clear out the persisted list no matter what, this will force
                 // rescanning of all plugin types including lazy ones.
                 // http://issues.umbraco.org/issue/U4-4789
-                File.Delete(pluginListFile);
+                if (File.Exists(pluginListFile))
+                    File.Delete(pluginListFile);
 
                 // always set to true if we're not detecting (generally only for testing)
                 RequiresRescanning = true;
@@ -180,23 +184,20 @@ namespace Umbraco.Core
         /// <summary>
         /// Gets the currently cached hash value of the scanned assemblies.
         /// </summary>
-        /// <value>The cached hash value, or 0 if no cache is found.</value>
-        internal long CachedAssembliesHash
+        /// <value>The cached hash value, or string.Empty if no cache is found.</value>
+        internal string CachedAssembliesHash
         {
             get
             {
-                if (_cachedAssembliesHash != -1)
+                if (_cachedAssembliesHash != null)
                     return _cachedAssembliesHash;
 
                 var filePath = GetPluginHashFilePath();
-                if (File.Exists(filePath) == false) return 0;
+                if (File.Exists(filePath) == false) return string.Empty;
 
                 var hash = File.ReadAllText(filePath, Encoding.UTF8);
 
-                long val;
-                if (long.TryParse(hash, out val) == false) return 0;
-
-                _cachedAssembliesHash = val;
+                _cachedAssembliesHash = hash;
                 return _cachedAssembliesHash;
             }
         }
@@ -205,11 +206,11 @@ namespace Umbraco.Core
         /// Gets the current assemblies hash based on creating a hash from the assemblies in various places.
         /// </summary>
         /// <value>The current hash.</value>
-        internal long CurrentAssembliesHash
+        internal string CurrentAssembliesHash
         {
             get
             {
-                if (_currentAssembliesHash != -1)
+                if (_currentAssembliesHash != null)
                     return _currentAssembliesHash;
 
                 _currentAssembliesHash = GetFileHash(new List<Tuple<FileSystemInfo, bool>>
@@ -234,6 +235,14 @@ namespace Umbraco.Core
         private void WriteCachePluginsHash()
         {
             var filePath = GetPluginHashFilePath();
+
+            // be absolutely sure the folder exists
+            var folder = Path.GetDirectoryName(filePath);
+            if (folder == null)
+                throw new InvalidOperationException("The folder could not be determined for the file " + filePath);
+            if (Directory.Exists(folder) == false)
+                Directory.CreateDirectory(folder);
+
             File.WriteAllText(filePath, CurrentAssembliesHash.ToString(), Encoding.UTF8);
         }
 
@@ -245,41 +254,40 @@ namespace Umbraco.Core
         /// <returns>The hash.</returns>
         /// <remarks>Each file is a tuple containing the FileInfo object and a boolean which indicates whether to hash the
         /// file properties (false) or the file contents (true).</remarks>
-        internal static long GetFileHash(IEnumerable<Tuple<FileSystemInfo, bool>> filesAndFolders, ProfilingLogger logger)
+        internal static string GetFileHash(IEnumerable<Tuple<FileSystemInfo, bool>> filesAndFolders, ProfilingLogger logger)
         {
             using (logger.TraceDuration<PluginManager>("Determining hash of code files on disk", "Hash determined"))
             {
-                var hashCombiner = new HashCodeCombiner();
-
                 // get the distinct file infos to hash
                 var uniqInfos = new HashSet<string>();
                 var uniqContent = new HashSet<string>();
-
-                foreach (var fileOrFolder in filesAndFolders)
+                using (var generator = new HashGenerator())
                 {
-                    var info = fileOrFolder.Item1;
-                    if (fileOrFolder.Item2)
+                    foreach (var fileOrFolder in filesAndFolders)
                     {
-                        // add each unique file's contents to the hash
-                        // normalize the content for cr/lf and case-sensitivity
-
-                        if (uniqContent.Contains(info.FullName)) continue;
-                        uniqContent.Add(info.FullName);
-                        if (File.Exists(info.FullName) == false) continue;
-                        var content = RemoveCrLf(File.ReadAllText(info.FullName));
-                        hashCombiner.AddCaseInsensitiveString(content);
+                        var info = fileOrFolder.Item1;
+                        if (fileOrFolder.Item2)
+                        {
+                            // add each unique file's contents to the hash
+                            // normalize the content for cr/lf and case-sensitivity
+                            if (uniqContent.Add(info.FullName))
+                            {
+                                if (File.Exists(info.FullName) == false) continue;
+                                var content = RemoveCrLf(File.ReadAllText(info.FullName));
+                                generator.AddCaseInsensitiveString(content);
+                            }
+                        }
+                        else
+                        {
+                            // add each unique folder/file to the hash
+                            if (uniqInfos.Add(info.FullName))
+                            {
+                                generator.AddFileSystemItem(info);
+                            }
+                        }
                     }
-                    else
-                    {
-                        // add each unique folder/file to the hash
-
-                        if (uniqInfos.Contains(info.FullName)) continue;
-                        uniqInfos.Add(info.FullName);
-                        hashCombiner.AddFileSystemItem(info);
-                    }
-                } 
-
-                return ConvertHashToInt64(hashCombiner.GetCombinedHashCode());
+                    return generator.GenerateHash();
+                }
             }
         }
 
@@ -303,38 +311,32 @@ namespace Umbraco.Core
         /// <param name="filesAndFolders">A collection of files.</param>
         /// <param name="logger">A profiling logger.</param>
         /// <returns>The hash.</returns>
-        internal static long GetFileHash(IEnumerable<FileSystemInfo> filesAndFolders, ProfilingLogger logger)
+        internal static string GetFileHash(IEnumerable<FileSystemInfo> filesAndFolders, ProfilingLogger logger)
         {
             using (logger.TraceDuration<PluginManager>("Determining hash of code files on disk", "Hash determined"))
             {
-                var hashCombiner = new HashCodeCombiner();
-
-                // get the distinct file infos to hash
-                var uniqInfos = new HashSet<string>();
-
-                foreach (var fileOrFolder in filesAndFolders)
+                using (var generator = new HashGenerator())
                 {
-                    if (uniqInfos.Contains(fileOrFolder.FullName)) continue;
-                    uniqInfos.Add(fileOrFolder.FullName);
-                    hashCombiner.AddFileSystemItem(fileOrFolder);
+                    // get the distinct file infos to hash
+                    var uniqInfos = new HashSet<string>();
+
+                    foreach (var fileOrFolder in filesAndFolders)
+                    {
+                        if (uniqInfos.Contains(fileOrFolder.FullName)) continue;
+                        uniqInfos.Add(fileOrFolder.FullName);
+                        generator.AddFileSystemItem(fileOrFolder);
+                    }
+                    return generator.GenerateHash();
                 }
-
-                return ConvertHashToInt64(hashCombiner.GetCombinedHashCode());
             }
-        }
-
-        /// <summary>
-        /// Converts a string hash value into an Int64.
-        /// </summary>
-        internal static long ConvertHashToInt64(string val)
-        {
-            long outVal;
-            return long.TryParse(val, NumberStyles.AllowHexSpecifier, CultureInfo.InvariantCulture, out outVal) ? outVal : 0;
         }
 
         #endregion
 
         #region Cache
+
+        private const int ListFileOpenReadTimeout = 4000; // milliseconds
+        private const int ListFileOpenWriteTimeout = 2000; // milliseconds
 
         /// <summary>
         /// Attemps to retrieve the list of types from the cache.
@@ -381,7 +383,7 @@ namespace Umbraco.Core
             if (File.Exists(filePath) == false)
                 return cache;
 
-            using (var stream = File.OpenRead(filePath))
+            using (var stream = GetFileStream(filePath, FileMode.Open, FileAccess.Read, FileShare.Read, ListFileOpenReadTimeout))
             using (var reader = new StreamReader(stream))
             {
                 while (true)
@@ -436,22 +438,57 @@ namespace Umbraco.Core
         }
 
         private string GetPluginListFilePath()
-        {
-            var filename = "umbraco-plugins." + NetworkHelper.FileSafeMachineName + ".list";
-            return Path.Combine(_tempFolder, filename);
+        {            
+            switch (GlobalSettings.LocalTempStorageLocation)
+            {                
+                case LocalTempStorage.AspNetTemp:
+                    return Path.Combine(HttpRuntime.CodegenDir, @"UmbracoData\umbraco-plugins.list");
+                case LocalTempStorage.EnvironmentTemp:
+                    var appDomainHash = HttpRuntime.AppDomainAppId.ToSHA1();
+                    var cachePath = Path.Combine(Environment.ExpandEnvironmentVariables("%temp%"), "UmbracoData",
+                        //include the appdomain hash is just a safety check, for example if a website is moved from worker A to worker B and then back
+                        // to worker A again, in theory the %temp%  folder should already be empty but we really want to make sure that its not
+                        // utilizing an old path
+                        appDomainHash);
+                    return Path.Combine(cachePath, "umbraco-plugins.list");
+                case LocalTempStorage.Default:                    
+                default:
+                    return Path.Combine(_tempFolder, "umbraco-plugins." + NetworkHelper.FileSafeMachineName + ".list");
+            }
         }
 
         private string GetPluginHashFilePath()
         {
-            var filename = "umbraco-plugins." + NetworkHelper.FileSafeMachineName + ".hash";
-            return Path.Combine(_tempFolder, filename);
+            switch (GlobalSettings.LocalTempStorageLocation)
+            {
+                case LocalTempStorage.AspNetTemp:
+                    return Path.Combine(HttpRuntime.CodegenDir, @"UmbracoData\umbraco-plugins.hash");
+                case LocalTempStorage.EnvironmentTemp:
+                    var appDomainHash = HttpRuntime.AppDomainAppId.ToSHA1();
+                    var cachePath = Path.Combine(Environment.ExpandEnvironmentVariables("%temp%"), "UmbracoData",
+                        //include the appdomain hash is just a safety check, for example if a website is moved from worker A to worker B and then back
+                        // to worker A again, in theory the %temp%  folder should already be empty but we really want to make sure that its not
+                        // utilizing an old path
+                        appDomainHash);
+                    return Path.Combine(cachePath, "umbraco-plugins.hash");
+                case LocalTempStorage.Default:
+                default:
+                    return Path.Combine(_tempFolder, "umbraco-plugins." + NetworkHelper.FileSafeMachineName + ".hash");
+            }
         }
 
         internal void WriteCache()
         {
             var filePath = GetPluginListFilePath();
 
-            using (var stream = File.Open(filePath, FileMode.Create, FileAccess.ReadWrite))
+            // be absolutely sure the folder exists
+            var folder = Path.GetDirectoryName(filePath);
+            if (folder == null)
+                throw new InvalidOperationException("The folder could not be determined for the file " + filePath);
+            if (Directory.Exists(folder) == false)
+                Directory.CreateDirectory(folder);
+            
+            using (var stream = GetFileStream(filePath, FileMode.Create, FileAccess.Write, FileShare.None, ListFileOpenWriteTimeout))
             using (var writer = new StreamWriter(stream))
             {
                 foreach (var typeList in _types.Values)
@@ -471,7 +508,28 @@ namespace Umbraco.Core
             // at the moment we write the cache to disk every time we update it. ideally we defer the writing
             // since all the updates are going to happen in a row when Umbraco starts. that being said, the
             // file is small enough, so it is not a priority.
-            WriteCache();            
+            WriteCache();
+        }
+
+        private static Stream GetFileStream(string path, FileMode fileMode, FileAccess fileAccess, FileShare fileShare, int timeoutMilliseconds)
+        {
+            const int pauseMilliseconds = 250;
+            var attempts = timeoutMilliseconds / pauseMilliseconds;
+            while (true)
+            {
+                try
+                {
+                    return new FileStream(path, fileMode, fileAccess, fileShare);
+                }
+                catch
+                {
+                    if (--attempts == 0)
+                        throw;
+
+                    LogHelper.Debug<PluginManager>(string.Format("Attempted to get filestream for file {0} failed, {1} attempts left, pausing for {2} milliseconds", path, attempts, pauseMilliseconds));
+                    Thread.Sleep(pauseMilliseconds);
+                }
+            }
         }
 
         #endregion
